[glyf] Minor typo change

+++ src/OT/glyf/SimpleGlyph.hh
@@ -129,7 +129,7 @@ struct SimpleGlyph
     unsigned count = points_.length;
     for (unsigned i = 0; i < count; i++)
     {
-      uint8_t flag = points_[i].flag;
+      unsigned flag = points_[i].flag;
       if (flag & short_flag)
       {
 	if (unlikely (p + 1 > end)) return false;
